Reverse order of transform matrix calculation
skimage.transform.warp wants the inverse transform, so that was directly
calculated with a reverse order of transforms and directions. That
anyway had to be inverted to calculate the resulting center and diameter
of the sun disk. This made the code confusing. Now the normal
transformation matrix is calculated in a sane way and then inverted for
skimage.transform.warp.