Fix bug where part of image might get clipped to black
If the sun disk is very close to the edge of the north or south edges of
the scan, the tilt correction might clip the disk because only the RA
axis of the scan is resized but not the Dec axis.

Now the other axis is correctly resized to keep sun inside the image.